inline InstallService implementation
