+++ src/test/java/mocks/TestInstallService.java
@@ -1,7 +1,7 @@
 package mocks;
 
 import engine.FormplayerConfigEngine;
-import services.impl.InstallServiceImpl;
+import services.InstallService;
 
 import java.io.File;
 import java.net.URL;
@@ -9,7 +9,7 @@ import java.net.URL;
 /**
  * Created by willpride on 12/7/16.
  */
-public class TestInstallService extends InstallServiceImpl {
+public class TestInstallService extends InstallService {
 
     @Override
     public FormplayerConfigEngine configureApplication(String reference) throws Exception {
